refactor(workspace): extract quspin-operator crate (PR 3/7)
Third of seven PRs splitting quspin-core. Severs the `operator → basis`
half of the cycle by moving operator type definitions (plus the
`Operator` and `ParseOp` traits) into a new quspin-operator crate that
has no basis dependency.

Kept in quspin-core (for now):
- `operator/*/dispatch.rs` — the `*OperatorInner` dispatch enums and
  their `apply_and_project_to`/`apply` methods. These depend on
  `basis::dispatch::SpaceInner` and will move to quspin-matrix in PR 5
  alongside `operator/apply.rs`.
- `operator/apply.rs` — the generic free function behind the dispatch
  wrappers. Also moves to quspin-matrix in PR 5.

`cargo check -p quspin-operator` now passes in isolation with zero
references to basis, qmatrix, hamiltonian, expm, krylov, or
linear_operator.

The `quspin_core::operator::*` public surface is unchanged — the
`*OperatorInner` enums are re-exported as before, and
`quspin_operator::*` is re-exported through `crate::operator::mod`.
quspin-py compiles untouched.
