+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "crates/quspin-types",
     "crates/quspin-bitbasis",
+    "crates/quspin-operator",
     "crates/quspin-core",
     "crates/quspin-py",
 ]
