+++ crates/quspin-core/src/operator/monomial/mod.rs
@@ -1,5 +1,5 @@
 pub mod dispatch;
-pub mod operator;
 
 pub use dispatch::MonomialOperatorInner;
-pub use operator::{MonomialOperator, MonomialTerm};
+pub use quspin_operator::monomial::operator;
+pub use quspin_operator::monomial::{MonomialOperator, MonomialTerm};
