+++ crates/quspin-core/src/operator/pauli/dispatch.rs
@@ -38,7 +38,7 @@ impl HardcoreOperatorInner {
         in_vec: &[num_complex::Complex<f64>],
         out_vec: &mut [num_complex::Complex<f64>],
         overwrite: bool,
-    ) -> Result<(), crate::error::QuSpinError> {
+    ) -> Result<(), quspin_types::QuSpinError> {
         match self {
             Self::Ham8(h) => super::super::apply::apply_and_project_to(
                 h, input, output, coeffs, in_vec, out_vec, overwrite,
@@ -56,7 +56,7 @@ impl HardcoreOperatorInner {
         in_vec: &[num_complex::Complex<f64>],
         out_vec: &mut [num_complex::Complex<f64>],
         overwrite: bool,
-    ) -> Result<(), crate::error::QuSpinError> {
+    ) -> Result<(), quspin_types::QuSpinError> {
         match self {
             Self::Ham8(h) => super::super::apply::apply_and_project_to(
                 h, space, space, coeffs, in_vec, out_vec, overwrite,
