+++ crates/quspin-operator/src/bond/mod.rs
@@ -1,0 +1,3 @@
+pub mod operator;
+
+pub use operator::{BondOperator, BondTerm};
