+++ crates/quspin-operator/src/boson/mod.rs
@@ -1,0 +1,3 @@
+pub mod operator;
+
+pub use operator::{BosonOp, BosonOpEntry, BosonOperator};
