+++ crates/quspin-operator/src/fermion/mod.rs
@@ -1,0 +1,3 @@
+pub mod operator;
+
+pub use operator::{FermionOp, FermionOpEntry, FermionOperator};
