+++ crates/quspin-operator/src/monomial/mod.rs
@@ -1,0 +1,3 @@
+pub mod operator;
+
+pub use operator::{MonomialOperator, MonomialTerm};
