+++ crates/quspin-operator/src/pauli/mod.rs
@@ -1,0 +1,3 @@
+pub mod operator;
+
+pub use operator::{HardcoreOp, HardcoreOperator, OpEntry};
