+++ crates/quspin-operator/src/pauli/operator.rs
@@ -1,5 +1,5 @@
-use crate::operator::Operator;
-use crate::operator::ParseOp;
+use crate::Operator;
+use crate::ParseOp;
 use num_complex::Complex;
 use quspin_bitbasis::BitInt;
 use quspin_types::QuSpinError;
@@ -245,7 +245,7 @@ impl<C: Copy + Ord> Operator<C> for HardcoreOperator<C> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::operator::pauli::{HardcoreOp, OpEntry};
+    use crate::pauli::{HardcoreOp, OpEntry};
 
     use num_complex::Complex;
     use smallvec::smallvec;
