+++ crates/quspin-operator/src/spin/mod.rs
@@ -1,0 +1,3 @@
+pub mod operator;
+
+pub use operator::{SpinOp, SpinOpEntry, SpinOperator};
